Further cleanup.
