+++ src/main/resources/remotefutures.conf
@@ -1,7 +1,7 @@
 general {
     future-timeout-in-ms = 5000
     # remote-execution-context-classname = "org.remotefutures.core.impl.local.LocalRunningRemoteExecutionContext"
-    remote-execution-context-classname = "org.remotefutures.core.impl.akka.pullingworker_final.PullingWorkerRemoteExecutionContext"
+    remote-execution-context-classname = "org.remotefutures.core.impl.akka.pullingworker.PullingWorkerRemoteExecutionContext"
 
     #
     # Distribution strategy
