--- src/main/scala/org/remotefutures/core/impl/akka/pullingworker/MainManyMasters.scala
@@ -1,20 +1,0 @@
-package org.remotefutures.core.impl.akka.pullingworker
-
-object MainManyMasters extends Startup {
-
-  def main(args: Array[String]): Unit = {
-    val joinAddress = startBackend(None, "backend-shard1")
-    Thread.sleep(5000)
-    startBackend(Some(joinAddress), "backend-shard1")
-    startWorker(joinAddress)
-    Thread.sleep(5000)
-    startFrontend(joinAddress)
-
-    startBackend(Some(joinAddress), "backend-shard2")
-    startBackend(Some(joinAddress), "backend-shard2")
-    startWorker(joinAddress)
-    startWorker(joinAddress)
-    startWorker(joinAddress)
-  }
-
-}
