+++ src/main/scala/org/remotefutures/core/impl/akka/pullingworker/MasterWorkerProtocol.scala
@@ -1,3 +1,6 @@
+/*
+ * Copyright (c) 2014 Martin Senne
+ */
 package org.remotefutures.core.impl.akka.pullingworker
 
 object MasterWorkerProtocol {
