+++ src/main/scala/org/remotefutures/core/impl/akka/pullingworker/PullingWorkerRemoteExecutionContext.scala
@@ -18,24 +18,16 @@ import scala.concurrent.duration._
 import akka.util.Timeout
 
 
-object PullingWorkerRemoteExecutionContext {
-  /**
-   * Execute case class. Used by CallerActor and CalleeActor.
-   * @param body is the code to execute on callee (remote node) site.
-   */
-  case class Execute[T](body: () => T)
-}
 
-class PullingWorkerRemoteExecutionContext(settings: Settings, reporter: Throwable => Unit)
+
+class PullingWorkerRemoteExecutionContext(settings: Settings, reporter: Throwable ⇒ Unit)
   extends RemoteExecutionContext
   with Startup {
-
-  import PullingWorkerRemoteExecutionContext.Execute
 
   // =====================================================
   // this is the code to setup other nodes
   // =====================================================
-  println("Starting up pulling worker cluster.")
+  println("Starting up pulling worker (final) cluster.")
   val joinAddress = startBackend(None, "backend")
   Thread.sleep(5000)
   // startBackend(Some(joinAddress), "backend")
@@ -43,7 +35,7 @@ class PullingWorkerRemoteExecutionContext(settings: Settings, reporter: Throwabl
   // =====================================================
 
   val frontendSetup: FrontendSetup = new FrontendSetup(joinAddress, systemName)
-  // TODO: This is really bloody. We need a mechanism to check, if cluster (master and worker nodes) is up.
+  // TODO: This is really bloody. We need a mechanism to check, if cluster (master and worker nodes) are up.
   Thread.sleep(5000)
 
   def rnd = ThreadLocalRandom.current
@@ -53,25 +45,39 @@ class PullingWorkerRemoteExecutionContext(settings: Settings, reporter: Throwabl
    * Execute a function in its given context on a distant place.
    *
    * @param body is the code to execute and return T eventually
-   * @param bodyContext is the context/closure of of function fnc: () => T
+   * @param bodyContext is the context/closure of of function fnc: () ⇒ T
    * @tparam C specifies the Context type
    * @tparam T specifies the return tyoe
    */
-  override def execute[C, T](body: () => T, bodyContext: C, promise: Promise[T]): Unit = {
+  override def execute[C, T](body: () ⇒ T, bodyContext: C, promise: Promise[T]): Unit = {
 
     import scala.concurrent.ExecutionContext.Implicits.global
 
     implicit val timeout = Timeout(5 seconds)
 
-    val msg = Execute( () => body )
-    val possibleResult: Future[T] = (frontendSetup.remoteProducerActor ? msg).asInstanceOf[Future[T]]
+    // construct the message
+    val msg = Execute( body )
 
-    // complete the promise
-    possibleResult.onComplete {
-      case x ⇒ {
-        promise.complete(x)
-      }
-    }
+    // determine system and mediator
+    val system = frontendSetup.system
+    val mediator = frontendSetup.mediator
+
+    // create fresh actor (which handles master ack and does retries)
+    val remoteProducerActor = system.actorOf( Props(classOf[RemoteProducerActor], mediator, promise))
+
+    println("Sending message to actor " + remoteProducerActor)
+
+    remoteProducerActor ! msg
+
+//    val result: Future[T] = (remoteProducerActor ? msg).asInstanceOf[Future[T]]
+//
+//    // complete the promise
+//    result.onComplete {
+//      case x ⇒ {
+//        println("THIS is the IMPORTANT MESSAGE: Got result " + x + " in execute(...).")
+//        promise.complete(x)
+//      }
+//    }
 
 
     // old 3 =====================================================
@@ -79,7 +85,7 @@ class PullingWorkerRemoteExecutionContext(settings: Settings, reporter: Throwabl
 //      Props(classOf[RemoteProducerActor], frontendSetup.mediator, promise))
 //
 //    println("Sending execute to actor " + remoteProducerActor)
-//    remoteProducerActor ! Execute( () => body )
+//    remoteProducerActor ! Execute( () ⇒ body )
 
 
     // old 2 =====================================================
@@ -157,7 +163,9 @@ class FrontendSetup(joinAddress: akka.actor.Address, systemName: String) {
   val system = ActorSystem(systemName)
   Cluster(system).join(joinAddress)
   val mediator = DistributedPubSubExtension(system).mediator
-  val remoteProducerActor = system.actorOf( Props(classOf[RemoteProducerActor2], mediator))
+  println("Frontend Setup finished. Mediator is " + mediator)
+
+//   val remoteProducerActor = system.actorOf( Props(classOf[RemoteProducerActor], mediator))
 
   //  val frontend = actorSystem.actorOf(Props[Frontend], "frontend")
   //  actorSystem.actorOf(Props(classOf[WorkProducer], frontend), "producer")
@@ -169,11 +177,11 @@ class FrontendSetup(joinAddress: akka.actor.Address, systemName: String) {
  * @param mediatorToMaster
  * @param promise
  */
-class RemoteProducerActor2(mediatorToMaster: ActorRef, promise: Promise[Any]) extends Actor with ActorLogging {
-  import org.remotefutures.core.impl.akka.pullingworker.PullingWorkerRemoteExecutionContext.Execute
+class RemoteProducerActor(mediatorToMaster: ActorRef, promise: Promise[Any]) extends Actor with ActorLogging {
 
-  // register this actor to listen to result topic messages
-  mediatorToMaster ! DistributedPubSubMediator.Subscribe(Master.ResultsTopic, self)
+  // not required anymore
+    // register this actor to listen to result topic messages
+    // mediatorToMaster ! DistributedPubSubMediator.Subscribe(Master.ResultsTopic, self)
 
   // private var workerStates = Map[String, Promise[T]]()
 
@@ -191,25 +199,32 @@ class RemoteProducerActor2(mediatorToMaster: ActorRef, promise: Promise[Any]) ex
   override def postRestart(reason: Throwable): Unit = ()
 
   def receive = {
-    case Execute(body) => {
-      log.info("Remote producer actor got Execute.")
-      val work = Work(nextWorkId(), rnd.nextInt()) // TODO: this is intermediate and needs to be adapted to Work( ..., body )
+    case Execute(job) ⇒ {
+      log.info("Remote producer actor got Execute message.")
+
+      val work = Work(nextWorkId(), job) // HERE
+
+//      val body: () => Any = work.job
+//      log.info("RemoteProducerActor has  {} [{}].", body.hashCode(), body.getClass)
+//      val result: Any = body.apply()
+//      log.info("RemoteProducerActor has result" + result + " with type " + result.getClass + " and hash " + result.hashCode())
+
       mediatorToMaster ! Send("/user/master/active", work, localAffinity = false)
       //      (mediatorForMaster ? Send("/user/master/active", work, localAffinity = false)) map {
-      //        case Master.Ack(_) => Frontend.Ok
-      //      } recover { case _ => Frontend.NotOk } pipeTo sender
+      //        case Master.Ack(_) ⇒ Frontend.Ok
+      //      } recover { case _ ⇒ Frontend.NotOk } pipeTo sender
       context.become(waitForMasterAck(work), discardOld = false)
     }
-    case _: DistributedPubSubMediator.SubscribeAck => {
+    case _: DistributedPubSubMediator.SubscribeAck ⇒ {
       log.info("SubscribeAck for actor " + self)
     }
   }
 
   def waitForMasterAck(work: Work): Actor.Receive = {
-    case Master.Ack =>
+    case Master.Ack ⇒
       context.become(waitForWorkResult, discardOld = false)
     // TODO: Error handling ( see FrontEnd.NotOk )
-    case WorkResult(workId, result) =>
+    case WorkResult(workId, result) ⇒
       log.info("Whoops. Received result before Master.Ack. Still consuming result: {}", result)
       promise.success(result)
       log.info("Stopping remote producer actor " + self)
@@ -218,86 +233,12 @@ class RemoteProducerActor2(mediatorToMaster: ActorRef, promise: Promise[Any]) ex
   }
 
   def waitForWorkResult: Actor.Receive = {
-    case Master.Ack =>
+    case Master.Ack ⇒
       log.info("And Whoops. Received master ack after work result.")
-    case WorkResult(workId, result) =>
+    case WorkResult(workId, result) ⇒
       log.info("Consuming result: {}", result)
       promise.success(result)
       log.info("Stopping remote producer actor " + self)
       context.stop(self)
   }
 }
-
-
-
-
-// old 3 =====================================================
-
-
-///**
-// *
-// * @param mediatorToMaster
-// * @param promise
-// */
-//class RemoteProducerActor(mediatorToMaster: ActorRef, promise: Promise[Any]) extends Actor with ActorLogging {
-//  import org.remotefutures.core.impl.akka.pullingworker.PullingWorkerRemoteExecutionContext.Execute
-//
-//  // register this actor to listen to result topic messages
-//  mediatorToMaster ! DistributedPubSubMediator.Subscribe(Master.ResultsTopic, self)
-//
-//  private var workerStates = Map[String, Promise[T]]()
-//
-//  def scheduler = context.system.scheduler
-//
-//  def rnd = ThreadLocalRandom.current
-//  def nextWorkId(): String = UUID.randomUUID().toString
-//
-//  //  override def preStart(): Unit =
-//  //  // waiting five whopping seconds before sending a tick???.
-//  //  // Is there a particular reason for using this strategy?
-//  //    scheduler.scheduleOnce(5.seconds, self, Tick)
-//
-//  // override postRestart so we don't call preStart and schedule a new Tick
-//  override def postRestart(reason: Throwable): Unit = ()
-//
-//  def receive = {
-//    case Execute(body) => {
-//      log.info("Remote producer actor got Execute.")
-//      val work = Work(nextWorkId(), rnd.nextInt()) // TODO: this is intermediate and needs to be adapted to Work( ..., body )
-//      mediatorToMaster ! Send("/user/master/active", work, localAffinity = false)
-////      (mediatorForMaster ? Send("/user/master/active", work, localAffinity = false)) map {
-////        case Master.Ack(_) => Frontend.Ok
-////      } recover { case _ => Frontend.NotOk } pipeTo sender
-//      context.become(waitForMasterAck(work), discardOld = false)
-//    }
-//    case _: DistributedPubSubMediator.SubscribeAck => {
-//      log.info("SubscribeAck for actor " + self)
-//    }
-//  }
-//
-//  def waitForMasterAck(work: Work): Actor.Receive = {
-//    case Master.Ack =>
-//      context.become(waitForWorkResult, discardOld = false)
-//      // TODO: Error handling ( see FrontEnd.NotOk )
-//    case WorkResult(workId, result) =>
-//      log.info("Whoops. Received result before Master.Ack. Still consuming result: {}", result)
-//      promise.success(result)
-//      log.info("Stopping remote producer actor " + self)
-//      // 1) context.stop(self)
-//      context.become(waitForWorkResult, discardOld = false)
-//  }
-//
-//  def waitForWorkResult: Actor.Receive = {
-//    case Master.Ack =>
-//      log.info("And Whoops. Received master ack after work result.")
-//    case WorkResult(workId, result) =>
-//      log.info("Consuming result: {}", result)
-//      promise.success(result)
-//      log.info("Stopping remote producer actor " + self)
-//      context.stop(self)
-//  }
-//}
-//
-
-
-
