--- src/main/scala/org/remotefutures/core/impl/akka/pullingworker/WorkConsumer.scala
@@ -1,17 +1,0 @@
-package org.remotefutures.core.impl.akka.pullingworker
-
-import akka.actor.{Actor, ActorLogging}
-import akka.contrib.pattern.{DistributedPubSubExtension, DistributedPubSubMediator}
-
-class WorkConsumer extends Actor with ActorLogging {
-
-  val mediator = DistributedPubSubExtension(context.system).mediator
-  mediator ! DistributedPubSubMediator.Subscribe(Master.ResultsTopic, self)
-
-  def receive = {
-    case _: DistributedPubSubMediator.SubscribeAck =>
-    case WorkResult(workId, result) =>
-      log.info("Consumed result: {}", result)
-  }
-
-}
