+++ src/main/scala/org/remotefutures/core/impl/akka/pullingworker/WorkExecutor.scala
@@ -1,6 +1,9 @@
+/*
+ * Copyright (c) 2014 Martin Senne
+ */
 package org.remotefutures.core.impl.akka.pullingworker
 
-import akka.actor.Actor
+import akka.actor.{ActorLogging, Actor}
 
 object WorkExecutor {
   case class WorkComplete(result: Any)
@@ -22,12 +25,23 @@ object WorkExecutor {
 // specific, fully typed, worker actors.
 // the later is how pickler generates type specific serializers
 //
-class WorkExecutor extends Actor {
+class WorkExecutor extends Actor with ActorLogging{
 
   def receive = {
-    case n: Int =>
-      val n2 = n * n
-      val result = s"$n * $n = $n2"
-      sender ! WorkExecutor.WorkComplete(result)
+    case body: (() => Any) ⇒
+
+      // use () to actually execute the job
+      // val result: Any = body.apply()
+
+      val result: Any = body.apply()
+      log.info("WorkExecutor got job {} [{}].", body.hashCode(), body.getClass)
+
+      log.info("WorkExecutor has result: " + result + " with type " + result.getClass + " and hash " + result.hashCode())
+      // sender ! WorkExecutor.WorkComplete( body.apply() )
+      sender ! WorkExecutor.WorkComplete( result )
+//    case n: Int ⇒
+//      val n2 = n * n
+//      val result = s"$n * $n = $n2"
+//      sender ! WorkExecutor.WorkComplete(result)
   }
 }
