+++ src/main/scala/org/remotefutures/core/impl/akka/pullingworker/Worker.scala
@@ -1,3 +1,6 @@
+/*
+ * Copyright (c) 2014 Martin Senne
+ */
 package org.remotefutures.core.impl.akka.pullingworker
 
 import java.util.UUID
@@ -68,17 +71,17 @@ class Worker(clusterClient: ActorRef, workExecutorProps: Props, registerInterval
   var currentWorkId: Option[String] = None
 
   def workId: String = currentWorkId match {
-    case Some(workId) => workId
-    case None         => throw new IllegalStateException("Not working")
+    case Some(workId) ⇒ workId
+    case None         ⇒ throw new IllegalStateException("Not working")
   }
 
   // handle errors according to strategy
   override def supervisorStrategy = OneForOneStrategy() {
-    // partial function [Throwable => Directive] (see SupervisorStrategy.Decider)
-    case _: ActorInitializationException => Stop
-    case _: DeathPactException           => Stop
-    case _: Exception =>
-      currentWorkId foreach { workId => sendToMaster(WorkFailure(workerId, workId)) }
+    // partial function [Throwable ⇒ Directive] (see SupervisorStrategy.Decider)
+    case _: ActorInitializationException ⇒ Stop
+    case _: DeathPactException           ⇒ Stop
+    case _: Exception ⇒
+      currentWorkId foreach { workId ⇒ sendToMaster(WorkFailure(workerId, workId)) }
       context.become(idle)
       Restart
   }
@@ -88,11 +91,12 @@ class Worker(clusterClient: ActorRef, workExecutorProps: Props, registerInterval
   def receive = idle
 
   def idle: Receive = {
-    case WorkNeedsToBeDone =>
+    case WorkNeedsToBeDone ⇒
       sendToMaster(RequestForWork(workerId))
 
-    case Work(workId, job) =>
-      log.debug("Got work: {}", job)
+    case Work(workId, job) ⇒
+      // log.info("Got work: {} with type {}", job, job.getClass)
+      log.info("Worker got work.")
       currentWorkId = Some(workId)
       // Marvin: While sending work for execution, how do you link the result back to the actual callback?
       // Is it done implicitly through actor reference? Just asking for clarification.
@@ -100,21 +104,21 @@ class Worker(clusterClient: ActorRef, workExecutorProps: Props, registerInterval
       // 1. delegate work to "exec" actor
       // 2. switch context ( see partial function PartialFunction[Any, Unit] "working" below ):
       // --> This actor (worker) will receive Worker.WorkComplete( result ) from "exec" actor. (
-      workExecutor ! job
+      workExecutor ! job // HERE
       context.become(working)
   }
 
   def working: Receive = {
     // work is completed on "workExecutor" actor
-    case WorkExecutor.WorkComplete(result) =>
-      log.debug("Work is complete. Result {}.", result)
+    case WorkExecutor.WorkComplete(result) ⇒
+      log.info("Work is complete. Result {}.", result)
       sendToMaster(WorkSuccess(workerId, workId, result))
       context.setReceiveTimeout(5.seconds)
       context.become(waitForWorkStatusAck(result))
 
     // Marvin: is queing an option or a bad idea in this case?
     // Martin: Which actor is queueing ? A worker pulls for work. (see waitForWorkIsDoneAck)
-    case _: Work =>
+    case _: Work ⇒
       log.info("Yikes. Master told me to do work, while I'm working.")
   }
 
@@ -129,19 +133,19 @@ class Worker(clusterClient: ActorRef, workExecutorProps: Props, registerInterval
   //   2. MasterWorkerProtocol.RequestForWork is sent to master (see below) to request more work
 
   def waitForWorkStatusAck(result: Any): Receive = {
-    case WorkStatusAck(id) if id == workId =>
+    case WorkStatusAck(id) if id == workId ⇒
       sendToMaster(RequestForWork(workerId))
       context.setReceiveTimeout(Duration.Undefined)
       context.become(idle)
-    case ReceiveTimeout =>
+    case ReceiveTimeout ⇒
       log.info("No ack from master, retrying")
       sendToMaster(WorkSuccess(workerId, workId, result))
   }
 
   override def unhandled(message: Any): Unit = message match {
-    case Terminated(`workExecutor`) => context.stop(self)
-    case WorkNeedsToBeDone          =>
-    case _                          => super.unhandled(message)
+    case Terminated(`workExecutor`) ⇒ context.stop(self)
+    case WorkNeedsToBeDone          ⇒
+    case _                          ⇒ super.unhandled(message)
   }
 
   def sendToMaster(msg: Any): Unit = {
